update url exp to 30Day

+++ src/main/java/com/aptech/mymusic/presentation/internalmodel/Resource.java
@@ -17,7 +17,7 @@ import java.util.concurrent.TimeUnit;
 
 public class Resource {
 
-    public static final int URL_EXP = (int) TimeUnit.DAYS.toMillis(30);
+    public static final long URL_EXP = TimeUnit.DAYS.toMillis(30);
 
     public static String getUrl(Path path, String name) {
         return getUrl(ResourceConfig.SYSTEM_RESOURCE_TYPE, path, name);
